// src/app/Models/vendor.ts
export class vendor {
  id: string;

  vendorId: string;
  vendorName: string;
  password: string;
  email: string;
  contactNumber: string;
  location: string;
  address: string;
  ccEmailIds: string;
  active: string;

  status: number;
  message: string;

  contactName: string;
  vendorNameSearch: string;

  fromDateTime: string;
  toDateTime: string;
  vendorPriority: string;
}